+++ src/zearch/spider/Crawler.java
@@ -46,7 +46,7 @@ public class Crawler extends Thread {
         String url = urlQueue.remove();
         Document doc = Scraper.getDocumentFromURL(url);
         Scraper.parseLinks(url, doc, urlQueue);
-        Map<String,Integer> textGrams = Scraper.parseTextGrams(doc);
-        Map<String,Integer> metaGrams = Scraper.parseMetaGrams(doc);
+        Map<String, Short> textGrams = Scraper.parseTextGrams(doc);
+        Map<String, Short> metaGrams = Scraper.parseMetaGrams(doc);
     }
 }
