+++ tst/zearch/gram/GrammifierTest.java
@@ -10,20 +10,20 @@ class GrammifierTest {
 
     @Test
     void grammifyTest() {
-        Map<String, Integer> grams = Grammifier.grammify("aBcdEf!gh,ab   cde f;");
-        assertEquals(grams.get("abc"), 1);
-        assertEquals(grams.get("bcd"), 1);
-        assertEquals(grams.get("cde"), 2);
-        assertEquals(grams.get("def"), 1);
-        assertEquals(grams.get("efg"), 1);
-        assertEquals(grams.get("fgh"), 1);
-        assertEquals(grams.get("gha"), 1);
-        assertEquals(grams.get("hab"), 1);
-        assertEquals(grams.get("ab "), 1);
-        assertEquals(grams.get("b c"), 1);
-        assertEquals(grams.get(" cd"), 1);
-        assertEquals(grams.get("de "), 1);
-        assertEquals(grams.get("e f"), 1);
+        Map<String, Short> grams = Grammifier.grammify("aBcdEf!gh,ab   cde f;");
+        assertEquals(grams.get("abc"), (short) 1);
+        assertEquals(grams.get("bcd"), (short) 1);
+        assertEquals(grams.get("cde"), (short) 2);
+        assertEquals(grams.get("def"), (short) 1);
+        assertEquals(grams.get("efg"), (short) 1);
+        assertEquals(grams.get("fgh"), (short) 1);
+        assertEquals(grams.get("gha"), (short) 1);
+        assertEquals(grams.get("hab"), (short) 1);
+        assertEquals(grams.get("ab "), (short) 1);
+        assertEquals(grams.get("b c"), (short) 1);
+        assertEquals(grams.get(" cd"), (short) 1);
+        assertEquals(grams.get("de "), (short) 1);
+        assertEquals(grams.get("e f"), (short) 1);
         assertEquals(grams.keySet().size(), 13);
     }
 }
